数据库规范&BUG修改

+++ known-parent/known-common/src/main/java/com/known/common/model/LoginLog.java
@@ -11,7 +11,8 @@ import java.util.Date;
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class LoginLog {
-    private Integer id;
+
+    private Long id;
 
     private String account;
